REC[HB]::Cherenkov:  change nphe from int to float
Former-commit-id: 2d1ca3029c092e2031956243a018efa2d4ba91d6

+++ common-tools/clas-reco/src/main/java/org/jlab/clas/detector/DetectorData.java
@@ -266,7 +266,7 @@ public class DetectorData {
            bank.setFloat("dphi", row, (float) c.getDeltaPhi());
            bank.setFloat("path", row, (float) 0.0);
            bank.setFloat("time", row, (float) c.getTime());
-           bank.setInt("nphe", row, (int) c.getEnergy());
+           bank.setFloat("nphe", row, (float) c.getEnergy());
            bank.setFloat("chi2", row, (float) 0.0);
        }
        return bank;
